--- packages/vuetify/src/components/VValidate/index.ts
@@ -1,1 +1,0 @@
-export { default as VValidate } from './VValidate'
