+++ packages/vuetify/src/components/VValidation/index.ts
@@ -1,0 +1,1 @@
+export { default as VValidation } from './VValidation'
