+++ packages/vuetify/src/components/index.ts
@@ -83,5 +83,5 @@ export * from './VTooltip'
 // export * from './VTreeview'
 // export * from './VVirtualScroll'
 // export * from './VWindow'
-export * from './VValidate'
+export * from './VValidation'
 export * from './transitions'
